chore(rollover): Rename rollover modules
We "fix" the module names:

- The module `current` corresponds to the deprecated protocol, so we
rename it to deprecated.

- The module `next` corresponds to the current protocol, so we rename
it to `current`.

We do this in a separate commit to make it easier to review.
