+++ xtra-libp2p-rollover/src/current.rs
@@ -2,4 +2,4 @@ pub mod maker;
 pub mod protocol;
 pub mod taker;
 
-pub const PROTOCOL: &str = "/itchysats/rollover/2.0.0";
+pub const PROTOCOL: &str = "/itchysats/rollover/3.0.0";
