+++ xtra-libp2p-rollover/src/deprecated.rs
@@ -1,0 +1,5 @@
+pub mod maker;
+pub mod protocol;
+pub mod taker;
+
+pub const PROTOCOL: &str = "/itchysats/rollover/2.0.0";
