--- xtra-libp2p-rollover/src/next.rs
@@ -1,5 +1,0 @@
-pub mod maker;
-pub mod protocol;
-pub mod taker;
-
-pub const PROTOCOL: &str = "/itchysats/rollover/3.0.0";
